+++ record/example/lib/main.dart
@@ -167,7 +167,7 @@ class _AudioRecorderState extends State<AudioRecorder> {
           print('${AudioEncoder.aacLc.name} supported: $isSupported');
         }
 
-        // final devs = await _audioRecorder.listInputDevices();
+        final devs = await _audioRecorder.listInputDevices();
 
         await _audioRecorder.start();