+++ record_macos/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0
-- feat: Add input devices listing.
-- feat: Add number of channels.
+* feat: Add input devices listing.
+* feat: Add number of channels.
+* core: Moved to AVCaptureSession for input device selection.
 
 ## 0.1.0
 * core: Initial release.